Move authorizeJwt from provider to validator
